wii-esc-ng: Branch for 3d mode
git-svn-id: https://wii-esc.googlecode.com/svn/branches/wii-esc-ng-3d@512 5c42c5a8-900f-5b5a-2f8c-d720ba63385b

+++ src/start.h
@@ -74,7 +74,7 @@ static uint8_t start() {
       start_timing += ((int16_t)est_comm_time - start_timing) >> 2;
       if (++good_com >= ENOUGH_GOODIES) {
         good_com = ENOUGH_GOODIES;
-        if ((start_timing <= RPM_TO_COMM_TIME(RPM_START_MIN_RPM) * 2)) {
+        if ((start_timing <= (int16_t)RPM_TO_COMM_TIME(RPM_START_MIN_RPM) * 2)) {
           next_comm_state();
           change_comm_state(pwr_stage.com_state);
           return START_RES_OK;
